requirements.txt
requests
pandas
python-dotenv
tqdm
faster-whisper
matplotlib
numpy
